fix: add head in demo page
