+++ src/assets/css/index.css
@@ -120,9 +120,4 @@
 
 .demoBookPage{
   width: 100%;
-
-  @media screen and (min-width: 1680px) {
-      max-width: 1600px;
-      margin: 0 auto;
-  }
 }
